Add files via upload
Customer churn accounts for the amount of paying customers who will no longer be using a service or continue being a customer of a certain business. Churn prediction is essential for any successful business. High customer churn can not only be expensive for a business, but it also demonstrates that customers may be dissatisfied with the business’s products or services. This will negatively impact future growth and revenue. So, it is important to try to minimize customer churn by understanding what variables impact this potential problem. If a business can identify the variables that determine whether or not a customer will continue using their products or services, they can minimize churn and maximize profits.

All businesses are impacted by customer churn. Banks, for example, want customers to continue using their services and be loyal to their comapny for as long as possible. For banks, understanding customer churn is especially important because it is a competitive market, built on trust and satisfaction. In order to gain customer's loyalty keep their business to , the bank must anticipate what might impact a customer to leave and join another bank. Conversely, they may want to understand or identify trends in what attributes long term loyal customers have as well. For example, understanding patterns by viewing operational data, like tenure, credit card usage, or even country of residence can help a business predict the probability of churn for each customer.

Link to dataset: https://www.kaggle.com/datasets/gauravtopre/bank-customer-churn-dataset